Clean up everything except for auto.
