+++ pyserini/encode/__init__.py
@@ -24,6 +24,7 @@ from ._auto import AutoQueryEncoder, AutoDocumentEncoder
 from ._bpr import BprQueryEncoder
 from ._cached_data import CachedDataQueryEncoder
 from ._cosdpr import CosDprEncoder, CosDprDocumentEncoder, CosDprQueryEncoder
+from ._dkrr import DkrrDprQueryEncoder
 from ._dpr import DprDocumentEncoder, DprQueryEncoder
 from ._openai import OpenAIDocumentEncoder, OpenAIQueryEncoder, OPENAI_API_RETRY_DELAY
 from ._slim import SlimQueryEncoder
