+++ pyserini/search/faiss/__main__.py
@@ -21,21 +21,15 @@ import numpy as np
 from tqdm import tqdm
 
 from pyserini.encode import QueryEncoder
-
-from pyserini.encode import AnceQueryEncoder
-from pyserini.encode import BprQueryEncoder
-from pyserini.encode import CosDprQueryEncoder
-from pyserini.encode import DprQueryEncoder
-from pyserini.encode import TctColBertQueryEncoder
-
+from pyserini.encode import AggretrieverQueryEncoder, AnceQueryEncoder, BprQueryEncoder, CosDprQueryEncoder, \
+    DkrrDprQueryEncoder, DprQueryEncoder, TctColBertQueryEncoder
 from pyserini.encode._pca import PcaEncoder
 from pyserini.output_writer import get_output_writer, OutputFormat
 from pyserini.query_iterator import get_query_iterator, TopicsFormat
-from pyserini.search.faiss._searcher import (AutoQueryEncoder, AggretrieverQueryEncoder, OpenAIQueryEncoder,
-                                             DkrrDprQueryEncoder, ClipQueryEncoder)
+from pyserini.search.faiss import FaissSearcher, BinaryDenseFaissSearcher
+from pyserini.search.faiss import DenseVectorAveragePrf, DenseVectorRocchioPrf, DenseVectorAncePrf
+from pyserini.search.faiss._searcher import AutoQueryEncoder, OpenAIQueryEncoder, ClipQueryEncoder
 from pyserini.search.lucene import LuceneSearcher
-from ._prf import DenseVectorAveragePrf, DenseVectorRocchioPrf, DenseVectorAncePrf
-from pyserini.search.faiss import FaissSearcher, BinaryDenseSearcher
 
 # Fixes this error: "OMP: Error #15: Initializing libomp.a, but found libomp.dylib already initialized."
 # https://stackoverflow.com/questions/53014306/error-15-initializing-libiomp5-dylib-but-found-libiomp5-dylib-already-initial
@@ -215,14 +209,14 @@ if __name__ == '__main__':
         # create searcher from index directory
         if args.searcher.lower() == 'bpr':
             kwargs = dict(binary_k=args.binary_hits, rerank=args.rerank)
-            searcher = BinaryDenseSearcher(args.index, query_encoder)
+            searcher = BinaryDenseFaissSearcher(args.index, query_encoder)
         else:
             searcher = FaissSearcher(args.index, query_encoder)
     else:
         # create searcher from prebuilt index name
         if args.searcher.lower() == 'bpr':
             kwargs = dict(binary_k=args.binary_hits, rerank=args.rerank)
-            searcher = BinaryDenseSearcher.from_prebuilt_index(args.index, query_encoder)
+            searcher = BinaryDenseFaissSearcher.from_prebuilt_index(args.index, query_encoder)
         else:
             searcher = FaissSearcher.from_prebuilt_index(args.index, query_encoder)
